[In Progress] Matrix Slugs (#60)
* Add slug to matrix model

* Replace references to collection handle with slug in navigation and api calls

* Make sure that a fieldset is assigned before reference a collections fields

* Ensure matrixpages use slug as reference as well

+++ app/Listeners/WhenFieldsetIsDetached.php
@@ -17,15 +17,17 @@ class WhenFieldsetIsDetached
      */
     public function handle(FieldsetDetached $event)
     {
-        $fields = $event->fieldset->fields;
-        $table  = $event->model->getTable();
+        if($event->fieldset) {
+            $fields = $event->fieldset->fields;
+            $table  = $event->model->getTable();
 
-        $columns = $fields->map(function($field) {
-            return $field->handle;
-        })->toArray();
+            $columns = $fields->map(function($field) {
+                return $field->handle;
+            })->toArray();
 
-        Schema::table($table, function ($blueprint) use ($columns) {
-            $blueprint->dropColumn($columns);
-        });
+            Schema::table($table, function ($blueprint) use ($columns) {
+                $blueprint->dropColumn($columns);
+            });
+        }  
     }
 }
